Add files via upload
Version 3 of my water monitor code. The main new feature is the ability for it to alert a Telegram Group that I am in when the water level falls below a certain value. This triggers twice when the value drops below a certain value. The alert only goes out a limited time per a set time limit. This is to avoid spam